Add files via upload
This was used to assess how each HLQ performs at predicting 0 class (propaganda) or 1 class (none). Weighed and Macro values provided. We used this to determine whether the answer to question of True or False is indicative of propaganda for example when Q20 is True it is indicative of propaganda conversely for Q258 should use the False.